9-fastapi/requirements.txt
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.0
flower==2.0.1
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
humanize==4.10.0
idna==3.10
kombu==5.4.1
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
